# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CropPilot AI farmer advisory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farms:
    get:
      operationId: listFarms
      tags: [farms]
      summary: List the farmer's farms
      responses:
        "200":
          description: Farm list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createFarm
      tags: [farms]
      summary: Create a farm profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmInput"
      responses:
        "201":
          description: Created farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /farms/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getFarm
      tags: [farms]
      summary: Get a farm
      responses:
        "200":
          description: Farm details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateFarm
      tags: [farms]
      summary: Update a farm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmUpdate"
      responses:
        "200":
          description: Updated farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteFarm
      tags: [farms]
      summary: Delete a farm
      responses:
        "204":
          description: Deleted farm
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /diagnoses:
    get:
      operationId: listDiagnoses
      tags: [diagnoses]
      summary: List saved diagnoses
      responses:
        "200":
          description: Diagnosis list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Diagnosis"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createDiagnosis
      tags: [diagnoses]
      summary: Analyze crop symptoms or an image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisInput"
      responses:
        "201":
          description: Diagnosis saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diagnosis"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /diagnoses/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getDiagnosis
      tags: [diagnoses]
      summary: Get a saved diagnosis
      responses:
        "200":
          description: Diagnosis details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diagnosis"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: AI service is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Farm:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        state: { type: ["string", "null"] }
        district: { type: ["string", "null"] }
        soilType: { type: ["string", "null"] }
        soilPh: { type: ["number", "null"] }
        organicMatter: { type: ["string", "null"] }
        irrigationMethod: { type: ["string", "null"] }
        farmSizeAcres: { type: ["number", "null"] }
        primaryCrops:
          type: array
          items: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, state, district, soilType, soilPh, organicMatter, irrigationMethod, farmSizeAcres, primaryCrops, notes, createdAt, updatedAt]
    FarmInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        state: { type: ["string", "null"], maxLength: 100 }
        district: { type: ["string", "null"], maxLength: 100 }
        soilType: { type: ["string", "null"], maxLength: 100 }
        soilPh: { type: ["number", "null"], minimum: 0, maximum: 14 }
        organicMatter: { type: ["string", "null"], maxLength: 80 }
        irrigationMethod: { type: ["string", "null"], maxLength: 100 }
        farmSizeAcres: { type: ["number", "null"], minimum: 0 }
        primaryCrops:
          type: array
          items: { type: string, maxLength: 80 }
        notes: { type: ["string", "null"], maxLength: 4000 }
      required: [name]
    FarmUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        state: { type: ["string", "null"], maxLength: 100 }
        district: { type: ["string", "null"], maxLength: 100 }
        soilType: { type: ["string", "null"], maxLength: 100 }
        soilPh: { type: ["number", "null"], minimum: 0, maximum: 14 }
        organicMatter: { type: ["string", "null"], maxLength: 80 }
        irrigationMethod: { type: ["string", "null"], maxLength: 100 }
        farmSizeAcres: { type: ["number", "null"], minimum: 0 }
        primaryCrops:
          type: array
          items: { type: string, maxLength: 80 }
        notes: { type: ["string", "null"], maxLength: 4000 }
    DiagnosisResult:
      type: object
      properties:
        diseaseName: { type: string }
        scientificName: { type: ["string", "null"] }
        confidenceScore: { type: integer, minimum: 0, maximum: 100 }
        severity: { type: string, enum: [low, medium, high, critical] }
        matchedSymptoms: { type: array, items: { type: string } }
        possibleCauses: { type: array, items: { type: string } }
        immediateActions: { type: array, items: { type: string } }
        treatmentOrganic: { type: array, items: { type: string } }
        treatmentChemical: { type: array, items: { type: string } }
        preventionTips: { type: array, items: { type: string } }
        whenToSeekHelp: { type: string }
        disclaimer: { type: string }
      required: [diseaseName, scientificName, confidenceScore, severity, matchedSymptoms, possibleCauses, immediateActions, treatmentOrganic, treatmentChemical, preventionTips, whenToSeekHelp, disclaimer]
    Diagnosis:
      type: object
      properties:
        id: { type: string, format: uuid }
        farmId: { type: ["string", "null"], format: uuid }
        farmName: { type: ["string", "null"] }
        cropName: { type: string }
        symptomsText: { type: ["string", "null"] }
        result: { $ref: "#/components/schemas/DiagnosisResult" }
        language: { type: string, enum: [en, hi] }
        createdAt: { type: string, format: date-time }
      required: [id, farmId, farmName, cropName, symptomsText, result, language, createdAt]
    DiagnosisInput:
      type: object
      properties:
        farmId: { type: ["string", "null"], format: uuid }
        cropName: { type: string, minLength: 1, maxLength: 100 }
        symptomsText: { type: ["string", "null"], maxLength: 4000 }
        imageData: { type: string, maxLength: 11000000, description: Optional data URL for a JPEG, PNG, or WebP photo, analyzed immediately and not retained. }
        growthStage: { type: ["string", "null"], maxLength: 100 }
        noticedWhen: { type: ["string", "null"], maxLength: 100 }
        notes: { type: ["string", "null"], maxLength: 2000 }
        language: { type: string, enum: [en, hi] }
      required: [cropName, language]
...